chore(wmg): refactor pipeline
