+++ backend/wmg/pipeline/summary_cubes/constants.py
@@ -1,0 +1,14 @@
+DIMENSION_NAME_MAP_CENSUS_TO_WMG = {
+    "tissue_ontology_term_id": "tissue_original_ontology_term_id",
+    "tissue_general_ontology_term_id": "tissue_ontology_term_id",
+}
+
+EXPRESSION_SUMMARY_CUBE_CREATED_FLAG = "expression_summary_cube_created"
+CELL_COUNTS_CUBE_CREATED_FLAG = "cell_counts_cube_created"
+EXPRESSION_SUMMARY_DEFAULT_CUBE_CREATED_FLAG = "expression_summary_default_cube_created"
+MARKER_GENES_CUBE_CREATED_FLAG = "marker_genes_cube_created"
+FILTER_RELATIONSHIPS_CREATED_FLAG = "filter_relationships_created"
+PRIMARY_FILTER_DIMENSIONS_CREATED_FLAG = "primary_filter_dimensions_created"
+DATASET_METADATA_CREATED_FLAG = "dataset_metadata_created"
+
+PIPELINE_STATE_FILENAME = "pipeline_state.json"
